+++ client/src/components/charts/VelocityTimeChart.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import ScatterChart from 'components/ScatterChart';
+import ScatterChart from 'components/v2/ScatterChart';
 import { PURPLE, GREY } from './colours';
 
 /**
