+++ client/src/components/v2/ContentPage.js
