+++ client/src/components/v2/DeleteModal.js
