+++ client/src/components/v2/DeleteModal.stories.js
