+++ client/src/components/v2/FontAwesomeIcon.js
