+++ client/src/components/v2/ScatterChart.stories.tsx
@@ -1,6 +1,6 @@
 import { BLUE, GREEN, GREY, PURPLE } from 'components/charts/colours';
 import React from 'react';
-import { createStory, StoryTemplate } from '../utils/stories';
+import { createStory, StoryTemplate } from '../../utils/stories';
 import ScatterChart, { AxisProps, DataProps } from './ScatterChart';
 
 export default {
