+++ client/src/components/v2/ScatterChart.tsx
