+++ client/src/components/v2/WidgetListGroupItem.tsx
