+++ client/src/views/v2/CameraSettingsView.js
@@ -1,4 +1,4 @@
-import ContentPage from 'components/ContentPage';
+import ContentPage from 'components/v2/ContentPage';
 import CameraRecordingContainer from 'components/v2/CameraRecordingContainer';
 import CameraSettings from 'components/v2/CameraSettings';
 import CameraStatus from 'components/v2/CameraStatus';
