+++ client/src/views/v2/PowerModelCalibrationView.js
@@ -1,6 +1,6 @@
 import React, { useState, useCallback, useRef } from 'react';
 import { Alert, Button, Card, Form, InputGroup } from 'react-bootstrap';
-import ContentPage from 'components/ContentPage';
+import ContentPage from 'components/v2/ContentPage';
 import { setCalibration, resetCalibration } from 'api/v2/powerModel';
 
 /**
