+++ client/src/views/v2/PowerModelView.js
@@ -1,6 +1,6 @@
 import React, { useState, useCallback } from 'react';
 import { Alert, Button, Card } from 'react-bootstrap';
-import ContentPage from 'components/ContentPage';
+import ContentPage from 'components/v2/ContentPage';
 import { usePowerModelState } from 'api/v2/powerModel';
 
 /**
