+++ client/src/views/v2/SensorStatusView.js
@@ -1,6 +1,6 @@
 import { useStatus } from 'api/v2/sensors';
-import ContentPage from 'components/ContentPage';
-import WidgetListGroupItem from 'components/WidgetListGroupItem';
+import ContentPage from 'components/v2/ContentPage';
+import WidgetListGroupItem from 'components/v2/WidgetListGroupItem';
 import React from 'react';
 import { Badge, ListGroup } from 'react-bootstrap';
